learned abit about Casting.
like you can push a double into an int my do (int)YourVariablesName; but you will lose some data as the Int cant hold ex 5.5 so its will only be to 5 then you type it out to the console.

but you can go from an int to a double with no problems.